edda: removed commented line and set Gentium as a font for stanzas
This is mostly a desperate move, because I cannot
make main font even smaller, and I want stanzas to use
font of at least the same size as normal text. 
Optima almost fits, but there are still a few long lines.
So I will try to use two serifs on one page (which is bad, I know) for some time
and, meanwhile, I will be trying to find a better solution.